materials-assignment1/function/requirements.txt
functions-framework==3.*
Pillow==9.4.*
